cmd: add daemon subcommands
This commit uses the `github.com/takama/daemon package` to add subcommands
to manage an Alice node as a daemon. The package is cross-platform and
easy to use. Remember to run `dep ensure` to install the package.

To install the daemon service, run this command from the directory that
contains the core configuration file for your Alice node (or use the
``--core-config` flag to specify the location of the config file):

    alice daemon install

To launch the daemon service, run:

    alice daemon start

Alternatively, you should be able to start and stop the service using
the commands from your operating system's service manager.

You will probably need root privilege to run these commands.