coin: add DB interface and inmem implementation
The idea is that it should be used by both State and Chain as the
underlying storage, similar to what Ethereum does.

Added a simple in-memory implementation. In the future I will add a
LevelDB implementation, but for now the in-memory one can be used to
implement the state and chain.

I will probably also add a Batch interface once it's needed.