ci: switch to golangci-lint
It should run a bit faster. It doesn't provide a flag to exclude a folder,
so I used a build tag instead. It ignores test files by default. You can add
this to a file if you want the linter to ignore it:

    //+build !lint

I like having to add that to a file explicitly, it makes you feel bad
about it.